Bug fixes
Bug 1: A struct is referenced in a different branch than the one it was defined in for certain gepInst instructions. This caused segfaults in the alias reasoning. Fixed by creating a function to search an entire program function for a certain value pointer or instruction.

Bug 2: addAlias method of DisjointedPVAliasSets did not handle case where both sets exist. This case is now explicitly handled. addAlias code also cleaner.

Additional test cases added